Vehicle: Move MockVehicle to its own directory
This is a preliminary change to consolidate vehicle characteristics into
one submodule. Right now this is a one-on-one move of the MockVehicle
into this module, and import updates/reordering in other files.
