+++ mission_basic.py
@@ -15,11 +15,11 @@ import dronekit
 # Ensure that we can import from the current directory as a package since 
 # running a Python script directly does not define the correct package
 from __init__ import __package__
+from geometry import Geometry
 from settings import Arguments
 from trajectory import Mission
-from trajectory.MockVehicle import MockVehicle
 from trajectory.Monitor import Monitor
-from geometry import Geometry
+from vehicle.MockVehicle import MockVehicle
 
 # Main mission program
 class Setup(object):
